ssidconfuser: properly simulate attack with beacon protection
Once a fake SSID is being advertised, remove the BIGTK. This is done to
simulate that an attacker cannot recompute the MMIE in the beacon after
changing the SSID.

Also add debug messages to make it explicit that the 'real AP' is still
sending the BIGTK to the client. Since the simulated attacker cannot
change these handshake messages.