ssidconfuser: verify reference beacon after connecting
Use a config to force the client to use passive scanning. This ensures
the client captured a reference "pre-auth" beacon before connecting.
Once the BIGTK in the 4-way handshake is received, the authenticity of
the captured pre-auth beacon can be checked. If this check fails,
disconnect from the network.

Note that Multi-Link Operation networks are not supported. In case the
AP changes the BIGTK between when the client captured the reference
beacon and the transmission of the BIGTK in the 4-way handshake, the
check will fail, and the client will disconnect. In this
proof-of-concept the BIP_GMAC algorithm is also not supported.